improving repo structure
